Ensure events() can be used with completions API
