+++ src/test/php/com/openai/unittest/EventStreamTest.class.php
@@ -1,47 +1,11 @@
 <?php namespace com\openai\unittest;
 
 use com\openai\rest\EventStream;
-use io\streams\{InputStream, MemoryInputStream};
 use lang\IllegalStateException;
 use test\{Assert, Test, Values};
 
 class EventStreamTest {
-
-  /** Streams contents */
-  private function contentStream(): array {
-    return [
-      'data: {"choices":[{"delta":{"role":"assistant"}}]}',
-      'data: {"choices":[{"delta":{"content":"Test"}}]}',
-      'data: {"choices":[{"delta":{"content":"ed"}}]}',
-      'data: [DONE]'
-    ];
-  }
-
-  /** Streams tool calls */
-  private function toolCallStream(): array {
-    return [
-      'data: {"choices":[{"delta":{"role":"assistant"}}]}',
-      'data: {"choices":[{"delta":{"tool_calls":[{"type":"function","function":{"name":"search","arguments":""}}]}}]}',
-      'data: {"choices":[{"delta":{"tool_calls":[{"function":{"arguments":"{"}}]}}]}',
-      'data: {"choices":[{"delta":{"tool_calls":[{"function":{"arguments":"}"}}]}}]}',
-      'data: {"choices":[{"delta":{},"finish_reason":"function_call"}]}',
-      'data: [DONE]'
-    ];
-  }
-
-  /** Returns input */
-  private function input(array $lines): InputStream {
-    return new MemoryInputStream(implode("\n\n", $lines));
-  }
-
-  /** Maps deltas to a list of pairs */
-  private function pairsOf(iterable $deltas): array {
-    $r= [];
-    foreach ($deltas as $field => $delta) {
-      $r[]= [$field => $delta];
-    }
-    return $r;
-  }
+  use Streams;
 
   /** Filtered deltas */
   private function filtered(): iterable {
@@ -71,13 +35,13 @@ class EventStreamTest {
   public function deltas() {
     Assert::equals(
       [['role' => 'assistant'], ['content' => 'Test'], ['content' => 'ed']],
-      $this->pairsOf((new EventStream($this->input($this->contentStream())))->deltas())
+      $this->pairsOf((new EventStream($this->input($this->contentCompletions())))->deltas())
     );
   }
 
   #[Test]
   public function deltas_throws_if_already_consumed() {
-    $events= new EventStream($this->input($this->contentStream()));
+    $events= new EventStream($this->input($this->contentCompletions()));
     iterator_count($events->deltas());
 
     Assert::throws(IllegalStateException::class, fn() => iterator_count($events->deltas()));
@@ -87,7 +51,7 @@ class EventStreamTest {
   public function ignores_newlines() {
     Assert::equals(
       [['role' => 'assistant'], ['content' => 'Test'], ['content' => 'ed']],
-      $this->pairsOf((new EventStream($this->input(['', ...$this->contentStream()])))->deltas())
+      $this->pairsOf((new EventStream($this->input(['', ...$this->contentCompletions()])))->deltas())
     );
   }
 
@@ -95,7 +59,7 @@ class EventStreamTest {
   public function filtered_deltas($filter, $expected) {
     Assert::equals(
       $expected,
-      $this->pairsOf((new EventStream($this->input($this->contentStream())))->deltas($filter))
+      $this->pairsOf((new EventStream($this->input($this->contentCompletions())))->deltas($filter))
     );
   }
 
@@ -103,7 +67,7 @@ class EventStreamTest {
   public function result() {
     Assert::equals(
       ['choices' => [['message' => ['role' => 'assistant', 'content' => 'Tested']]]],
-      (new EventStream($this->input($this->contentStream())))->result()
+      (new EventStream($this->input($this->contentCompletions())))->result()
     );
   }
 
@@ -116,7 +80,7 @@ class EventStreamTest {
         ['tool_calls' => [['function' => ['arguments' => '{']]]],
         ['tool_calls' => [['function' => ['arguments' => '}']]]],
       ],
-      $this->pairsOf((new EventStream($this->input($this->toolCallStream())))->deltas())
+      $this->pairsOf((new EventStream($this->input($this->toolCallCompletions())))->deltas())
     );
   }
 
@@ -128,7 +92,7 @@ class EventStreamTest {
         'message'       => ['role' => 'assistant', 'tool_calls' => $calls],
         'finish_reason' => 'function_call',
       ]]],
-      (new EventStream($this->input($this->toolCallStream())))->result()
+      (new EventStream($this->input($this->toolCallCompletions())))->result()
     );
   }
 }
